+++ feature-user-impl/src/main/java/com/nefrit/users/UsersRouter.kt
@@ -2,7 +2,7 @@ package com.nefrit.users
 
 interface UsersRouter {
 
-    fun openUser(userId: Int)
+    fun openUser(userId: Long)
 
     fun returnToUsers()
 }
