+++ feature-user-impl/src/main/java/com/nefrit/users/presentation/details/di/UserComponent.kt
@@ -19,7 +19,7 @@ interface UserComponent {
 
         fun create(
             @BindsInstance fragment: Fragment,
-            @BindsInstance userId: Int
+            @BindsInstance userId: Long
         ): UserComponent
     }
